cassandra-multiDCs/webserver/requirements.txt
requests
flask
cassandra-driver
matplotlib
pandas